break out the contract unit test into smaller unit tests
