+++ tests/integration/contracts/testdata/function_literals.go
@@ -1,0 +1,6 @@
+package testdata
+
+func frr() {
+	ffA := func(a int) int { return 0 }
+	ffB := ffA(2)
+}
