+++ tests/integration/contracts/testdata/general.go
@@ -1,0 +1,14 @@
+package testdata
+
+// Test literal contract
+var c = 2
+
+// Test typed variable
+var d int = 3
+
+// Test binary contract
+var e = 2 + 1
+
+var f = d + 1
+
+var ff = e + f
