+++ tests/integration/contracts/testdata/pointers.go
@@ -1,0 +1,8 @@
+package contracts
+
+func f() string {
+	a := "ahoj"
+	ra := &a
+	// pointers
+	da := *ra
+}
