+++ tests/integration/contracts/testdata/type_casting.go
@@ -1,0 +1,9 @@
+package testdata
+
+type Int int
+
+func f() {
+	// type assertion
+	asA := Int(1)
+	asB := asA.(int)
+}
